+++ src/Adsefid.Sdk/Sms/Models/SendSingleSms.cs
@@ -53,7 +53,7 @@ public sealed class SendSingleSmsResponse
     public LineSelector LineSelector { get; init; }
 
     [JsonPropertyName("cost")]
-    public required long Cost { get; init; }
+    public required decimal Cost { get; init; }
 
     [JsonPropertyName("receptor")]
     public required string Receptor { get; init; }
